refactor: setup colors
