+++ package.json
@@ -29,7 +29,7 @@
     "eslint": "^9.37.0",
     "eslint-config-next": "15.5.4",
     "postcss": "^8.5.6",
-    "tailwindcss": "^4",
+    "tailwindcss": "^4.14",
     "typescript": "^5.9.3"
   }
 }
