+++ src/app/globals.css
@@ -1,1 +1,11 @@
 @import "tailwindcss";
+
+@custom-variant dark (&:where(.dark, .dark *));
+
+@theme {
+  --color-primary: var(--color-blue-600);
+  --color-black: var(--color-neutral-950);
+  --color-gary-primary: var(--color-neutral-800);
+  --color-gray-secondary: var(--color-neutral-700);
+  --color-gray-light: var(--color-neutral-100);
+}
