+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css';
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
-import { ThemeProvider } from './theme-provider';
+import { ThemeProvider } from './themeProvider';
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
